01 - Initial CreateAccount imlementation on backend - not tested

+++ 01-basic-web-app/backend/domain/src/main/java/org/digitalcrafting/eregold/domain/accounts/AccountModel.java
@@ -3,11 +3,13 @@ package org.digitalcrafting.eregold.domain.accounts;
 import lombok.Builder;
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 @Builder
 public class AccountModel {
     private String accountNumber;
-    private Double currentBalance;
+    private BigDecimal currentBalance;
     private String currency;
     private AccountTypeEnum type;
 }
